separated generateRequest from MainActivity
